VanillaSAE/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
wandb>=0.15.0
h5py>=3.7.0
anndata>=0.8.0
scanpy>=1.9.0
tqdm>=4.62.0
PyYAML>=6.0
tensorboard>=2.8.0